requirements.txt
django==1.11.13
django-filter==1.1.0
djangorestframework==3.7.1
httpie